Clase 06/codigo/requirements.txt
python-dotenv
httpx
openai-agents
